+++ examples/src/bin/gl-interop.rs
@@ -26,7 +26,10 @@ mod linux {
         format::Format,
         image::{view::ImageView, ImageCreateFlags, ImageUsage, StorageImage, SwapchainImage},
         impl_vertex,
-        instance::{debug::DebugCallback, Instance, InstanceCreateInfo, InstanceExtensions},
+        instance::{
+            debug::{DebugUtilsMessenger, DebugUtilsMessengerCreateInfo},
+            Instance, InstanceCreateInfo, InstanceExtensions,
+        },
         pipeline::{
             graphics::{
                 color_blend::ColorBlendState,
@@ -391,16 +394,21 @@ mod linux {
         })
         .unwrap();
 
-        let _debug_callback = DebugCallback::errors_and_warnings(&instance, |msg| {
-            println!(
-                "{} {:?} {:?}: {}",
-                msg.layer_prefix.unwrap_or("unknown"),
-                msg.ty,
-                msg.severity,
-                msg.description
-            );
-        })
-        .unwrap();
+        let _debug_callback = unsafe {
+            DebugUtilsMessenger::new(
+                instance.clone(),
+                DebugUtilsMessengerCreateInfo::user_callback(Arc::new(|msg| {
+                    println!(
+                        "{} {:?} {:?}: {}",
+                        msg.layer_prefix.unwrap_or("unknown"),
+                        msg.ty,
+                        msg.severity,
+                        msg.description
+                    );
+                })),
+            )
+            .unwrap()
+        };
 
         let event_loop = EventLoop::new();
         let surface = WindowBuilder::new()